fixed initial bugs and finished valid_args()
Changelog:
*files now compile.
*valid_args(), tests if files are openable.
Reads, parse them into array of queue of vector of strings.
Thus, on runtime, each ATM will read a "line" (vector), like reading a file.
After cmd was used, extracted from the queue.
The vector, is a split version of each cmd line. Separated by spaces.
*Fixed compilation errors and incorrect code. Like get() function with const, while we call a Lock and Unlock functions within.
*Added split() function.
*Compiler made me create override to the default constructor.
*emplace is probably a righter way of use for map. As Account object is a mutex wrapper. Anyway maybe the override default constructor solved the bug.
Or maybe the change to emplace.
As the latter, instead of creating a new object from copy.
It's create in place of each slot.
Which is probably correct as mutexes aren't copy able or moveable.